api/requirements.txt
fastapi
uvicorn
python-multipart
onnxruntime
numpy
pillow
groq
python-dotenv
aiogram
